+++ src/expert_train.py
@@ -6,19 +6,7 @@ import torch.optim as optim
 from torch.utils.data import TensorDataset, DataLoader
 
 from simrocketenv import SimRocketEnv
-
-class ImitationNetwork(nn.Module):
-    def __init__(self, input_size, output_size):
-        super(ImitationNetwork, self).__init__()
-        self.fc1 = nn.Linear(input_size, 128)  # First MLP layer
-        self.fc2 = nn.Linear(128, 128)         # Second MLP layer
-        self.fc3 = nn.Linear(128, output_size) # Output layer
-
-    def forward(self, x):
-        x = torch.relu(self.fc1(x))
-        x = torch.relu(self.fc2(x))
-        x = self.fc3(x)
-        return x
+from nnpolicynetwork import NNPolicyNetwork
 
 def train_and_evaluate():
 
@@ -40,7 +28,7 @@ def train_and_evaluate():
     data_loader = DataLoader(dataset, batch_size=64, shuffle=True)
     input_size = observations.shape[1]
     output_size = actions.shape[1]
-    model = ImitationNetwork(input_size, output_size).to(device)
+    model = NNPolicyNetwork(input_size, output_size).to(device)
     optimizer = optim.Adam(model.parameters(), lr=0.001)
     criterion = nn.MSELoss()
